feat: setup Play Store release - rename package to com.eddiejdi.estouaqui
- Rename package from com.example.estou_aqui to com.eddiejdi.estouaqui
- Add release signing config with upload keystore
- Update google-services.json with new package name
- Move MainActivity.kt to new package directory
- Add keystore files to .gitignore for security
- Build config ready for Play Store AAB release
